+++ dapp/src/sendEmail.js
@@ -39,6 +39,7 @@ async function start() {
     emailSubject: process.env.IEXEC_REQUESTER_SECRET_1,
     emailContent: process.env.IEXEC_REQUESTER_SECRET_2,
     contentType: options.contentType,
+    senderName: options.senderName,
   };
   const envVars = validateInputs(unsafeEnvVars);
   const email = await extractZipAndBuildJson(
@@ -55,6 +56,7 @@ async function start() {
     emailContent: envVars.emailContent,
     mailJetSender: envVars.mailJetSender,
     contentType: envVars.contentType,
+    senderName: envVars.senderName,
   });
 
   await writeTaskOutput(