+++ tests/e2e/sendEmail.test.ts
@@ -1,6 +1,7 @@
 import {
   IExecDataProtector,
   ProtectedDataWithSecretProps,
+  getWeb3Provider as dataprotectorGetWeb3Provider,
 } from '@iexec/dataprotector';
 import { beforeAll, describe, expect, it } from '@jest/globals';
 import { Wallet } from 'ethers';
@@ -20,7 +21,7 @@ describe('web3mail.sendEmail()', () => {
     providerWallet = getRandomWallet();
     consumerWallet = getRandomWallet();
     dataProtector = new IExecDataProtector(
-      getWeb3Provider(providerWallet.privateKey)
+      dataprotectorGetWeb3Provider(providerWallet.privateKey)
     );
     web3mail = new IExecWeb3mail(getWeb3Provider(consumerWallet.privateKey));
 
@@ -112,6 +113,51 @@ describe('web3mail.sendEmail()', () => {
       await sleep(5_000);
       await expect(web3mail.sendEmail(params)).rejects.toThrow(
         'Dataset order not found'
+      );
+    },
+    3 * MAX_EXPECTED_BLOCKTIME
+  );
+  it(
+    'should successfully send email with a valid senderName',
+    async () => {
+      const params = {
+        emailSubject: 'e2e mail object for test',
+        emailContent: 'e2e mail content for test',
+        protectedData: validProtectedData.address,
+        senderName: 'Product Team'
+      };
+
+      const sendEmailResponse = await web3mail.sendEmail(params);
+      expect(sendEmailResponse.taskId).toBeDefined();
+    },
+    3 * MAX_EXPECTED_BLOCKTIME
+  );
+  it(
+    'should fail to send email with an invalid (too short) senderName',
+    async () => {
+      const params = {
+        emailSubject: 'e2e mail object for test',
+        emailContent: 'e2e mail content for test',
+        protectedData: validProtectedData.address,
+        senderName: 'AB'
+      };
+      await expect(web3mail.sendEmail(params)).rejects.toThrow(
+        'senderName must be at least 3 characters'
+      );
+    },
+    3 * MAX_EXPECTED_BLOCKTIME
+  );
+  it(
+    'should fail to send email with an invalid (too long) senderName',
+    async () => {
+      const params = {
+        emailSubject: 'e2e mail object for test',
+        emailContent: 'e2e mail content for test',
+        protectedData: validProtectedData.address,
+        senderName: 'A very long sender name'
+      };
+      await expect(web3mail.sendEmail(params)).rejects.toThrow(
+        'senderName must be at most 20 characters'
       );
     },
     3 * MAX_EXPECTED_BLOCKTIME